{"schema":"decodex/commit/1","summary":"Split manual test support modules","authority":"manual"}
